[WEB-4097][WEB-3954] - Remove Add User button and remove extra email send
