+++ app/political-terms/components/LayoutWithAlphabet.js
@@ -1,40 +1,12 @@
 import Link from 'next/link'
-
-export const alphabet = [
-  'a',
-  'b',
-  'c',
-  'd',
-  'e',
-  'f',
-  'g',
-  'h',
-  'i',
-  'j',
-  'k',
-  'l',
-  'm',
-  'n',
-  'o',
-  'p',
-  'q',
-  'r',
-  's',
-  't',
-  'u',
-  'v',
-  'w',
-  'x',
-  'y',
-  'z',
-]
+import { ALPHABET } from '@shared/utils/alphabet'
 
 export default function LayoutWithAlphabet({ activeLetter = 'A', children }) {
   return (
     <div className="flex lg:block mt-4 lg:mt-16">
       <div className="text-zinc-500">
         <div className="mb-1 lg:mb-0 lg:inline-block lg:mr-3">#</div>
-        {alphabet.map((letter) => (
+        {ALPHABET.map((letter) => (
           <Link
             key={letter}
             href={`/political-terms/${letter.toLowerCase()}`}
